Query based pagination
Queries records with
 - A page limit
 - Cursor (lastEvaluatedKey)
Replaces `ascending` and `descending` with `Direction` enum to sort the query results
Rename `order` to `orderBy`

Cursor tells the last processed item and this will be varied based on applied filters.
Masks the cursor to hide the actual db implementation details (partition key and range key).
Uses `crypto-js` to encrypt and decrypt the cursor.